requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
pydantic==2.5.0
aiohttp==3.9.1
pandas==2.1.4
openpyxl==3.1.2
httpx==0.25.2
fake-useragent==1.4.0
lxml==4.9.3
tqdm==4.66.1
deepseek-api==1.0.0
streamlit>=1.28.0
